I created a SignUp screen with message in new floder 'Components'in'src',also i moved sign-up form to separete component.I created a new files SignUpForm.js and SignInForm.js where he created a new component for next task
