+++ src/Components/Footer_component.js
@@ -1,5 +1,5 @@
 import "../App.css";
-import React, {useState} from "react";
+import React from "react";
 
 function Footer (){
         return (
