+++ src/Components/WelcomeSignInScreen.js
@@ -1,0 +1,13 @@
+import "../App.css";
+import React from "react";
+function WelcomeSignInScreen (){
+    return(
+        <div className="home-text-header" >
+            <h1>
+                Here may be your profile
+            </h1>
+        </div>
+    )
+};
+export default WelcomeSignInScreen;
+
