+++ src/Components/WelcomeSignUpScreen.js
@@ -1,0 +1,18 @@
+import "../App.css";
+import React from "react";
+
+function WelcomeSignUpScreen() {
+    return (
+        <div className="home-text-header">
+            <p> Thank you for filling out our sign up form.</p>
+            <p>We are glad that you joined us.</p>
+            <p> For this reason, we are
+                giving you a special offer.</p>
+            <p> Here is the link to the coupon for you to exclusively use on our website.</p>
+        </div>
+    )
+};
+
+
+export default WelcomeSignUpScreen;
+
