+++ src/Pages/sign_in.js
@@ -1,37 +1,13 @@
 import '../App.css';
-import Button from "@material-ui/core/Button";
-import TextField from "@material-ui/core/TextField";
-import axios from 'axios'
 import Footer from "../Components/Footer_component";
-import React from "react";
+import React, {useState} from "react";
+import WelcomeSignInScreen from "../Components/WelcomeSignUpScreen";
+import SignInForm from "../Components/SignUpForm";
 
 
 function Login() {
-    let email = '';
-    let password = '';
-    const handleChangeEmail = (event) => {
-        email = event.target.value;
-    }
-    const handleChangePassword = (event) => {
-        password = event.target.value
-    }
-    const handleChange = async () => {
-        if (!password || !email) {
-            /**
-             * TODO:
-             * Need to add some user friendly message if validation is not path
-             */
-            alert ("Not validation password or email,please try again")
-        }
-        await axios({
-            url: 'http://localhost:3000/auth/login',
-            data: {
-                email: email,
-                password: password
-            },
-            method: 'post'
-        },).then(rs => console.log(rs)).catch(e => console.log('error :', e))
-    }
+    const [isSignIn, setSignInStatus] = useState(false);
+
     return (
         <div className="App">
             <header className="App-header">
@@ -40,35 +16,7 @@ function Login() {
                         Website
                     </a>
                 </h1>
-                <div className="main-input-container" >
-                    <div className="input-container"  >
-                        <TextField
-                            id="outlined-basic1"
-                            label="Email"
-                            type="email"
-                            placeholder="Email"
-                            variant="outlined"
-                            name="emailInput"
-                            onChange={(e) => handleChangeEmail(e)}
-                        />
-                    </div>
-                    <div className="input-container">
-                        <TextField
-                            id="outlined-basic2"
-                            label="Password"
-                            variant="outlined"
-                            type="password"
-                            placeholder="Email"
-                            name="emailInput"
-                            onChange={(e) => handleChangePassword(e)}
-                        />
-                    </div>
-                    <div className="button-container">
-                        <Button  type="submit" onClick={handleChange} variant="contained" color="primary" href="#contained-buttons" >
-                            Send
-                        </Button>
-                    </div>
-                </div>
+                {isSignIn ? <WelcomeSignInScreen/> : <SignInForm  signInHandler={setSignInStatus}/>}
                 <div className="App-Footer">
                     <Footer/>
                 </div>
@@ -76,4 +24,5 @@ function Login() {
         </div>
     );
 }
+
 export default Login;
